+++ example/lib/pages/editing_page.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:dart_flow/dart_flow.dart';
+import 'package:node_flow/node_flow.dart';
 
 import '../demo_node.dart';
 
@@ -71,7 +71,7 @@ class _EditingPageState extends State<EditingPage> {
       ),
       body: Stack(
         children: <Widget>[
-          DartFlow<DemoNode, Object?>(
+          NodeFlow<DemoNode, Object?>(
             controller: _controller,
             nodeBuilder: (context, node) => DemoNodeCard(node: node),
           ),