+++ example/lib/pages/static_page.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:dart_flow/dart_flow.dart';
+import 'package:node_flow/node_flow.dart';
 
 import '../demo_node.dart';
 
@@ -73,7 +73,7 @@ class _StaticPageState extends State<StaticPage> {
           ),
         ],
       ),
-      body: DartFlow<DemoNode, Object?>(
+      body: NodeFlow<DemoNode, Object?>(
         controller: _controller,
         nodeBuilder: (context, node) => DemoNodeCard(node: node),
       ),