+++ test/controller_test.dart
@@ -1,7 +1,7 @@
 import 'dart:ui';
 
 import 'package:flutter_test/flutter_test.dart';
-import 'package:dart_flow/dart_flow.dart';
+import 'package:node_flow/node_flow.dart';
 
 FlowNode<String> node(String id, double x, double y, {Size? size}) => FlowNode(
   id: id,